store images in db

+++ Server/Services/RecipeService.cs
@@ -15,11 +15,13 @@ namespace AmnasKitchen.Server.Services
     {
         private readonly DatabaseConnectionHandler _databaseConnectionHandler;
         private readonly IMemoryCache _memoryCache;
+        private readonly IAkImageFileService akImageFileService;
 
-        public RecipeService(IMemoryCache memoryCache ,DatabaseConnectionHandler connectionHandler)
+        public RecipeService(IMemoryCache memoryCache, DatabaseConnectionHandler connectionHandler, IAkImageFileService akImageFileService)
         {
             _databaseConnectionHandler = connectionHandler ?? throw new ArgumentNullException(nameof(connectionHandler));
             _memoryCache = memoryCache ?? throw new ArgumentNullException(nameof(memoryCache));
+            this.akImageFileService = akImageFileService ?? throw new ArgumentNullException(nameof(akImageFileService));
         }
 
         public async Task<Recipe> GetRecipeById(int recipeId)
@@ -48,9 +50,20 @@ namespace AmnasKitchen.Server.Services
             {
                 using var connection = _databaseConnectionHandler.GetDbConnection();
                 connection.Open();
+                var sql = "Select c.*, ui.* from sa_amna.Categories c inner join sa_amna.UploadedImages ui on c.ImageId = ui.Id";
 
-                var result = await connection.QueryAsync<Category>("Select * from sa_amna.Categories");
+                var result = await connection.QueryAsync<Category, UploadedImage, Category>(
+                                 sql,
+                                 map: (c, u) =>
+                                     {
+                                         c.UploadedImage = u;
+
+                                         return c;
+                                     },
+                                 splitOn: "ImageId");
+
                 _memoryCache.Set(CacheKeys.AllCategories, result, DateTimeOffset.UtcNow.AddHours(2));
+
                 return result;
             }
         }
@@ -64,25 +77,39 @@ namespace AmnasKitchen.Server.Services
 
             using var connection = _databaseConnectionHandler.GetDbConnection();
             connection.Open();
-
             var sql = "INSERT INTO sa_amna.Recipe (CategoryId, Time, TimeUnit, Difficulty, Serving) Values (@categoryId, @time, @timeUnit, @difficulty, @serving)";
-            
-            await connection.ExecuteAsync(sql, new {recipe.CategoryId, recipe.Time, recipe.TimeUnit, recipe.Difficulty, recipe.Serving});
+
+            await connection.ExecuteAsync(
+                sql,
+                new
+                    {
+                        recipe.CategoryId,
+                        recipe.Time,
+                        recipe.TimeUnit,
+                        recipe.Difficulty,
+                        recipe.Serving
+                    });
         }
 
-        public async Task CreateCategory(Category category)
+        public async Task CreateCategory(CreateCategoryFormData categoryFormData)
         {
-            if (category == null)
+            if (categoryFormData == null)
             {
-                throw new ArgumentNullException(nameof(category));
+                throw new ArgumentNullException(nameof(categoryFormData));
             }
 
             using var connection = _databaseConnectionHandler.GetDbConnection();
             connection.Open();
+            int? insertedImageId = null;
 
-            var sql = "INSERT INTO sa_amna.Categories (Name, Description, ImageUrl) VALUES(@name, @description, @imageUrl)";
+            if (!string.IsNullOrEmpty(categoryFormData.IconPath))
+            {
+                var imageData = await akImageFileService.GetFileAsBytes(categoryFormData.IconPath);
+                insertedImageId = await StoreImageInDb(imageData);
+            }
 
-            await connection.ExecuteAsync(sql, new { category.Name, category.Description , category.ImageUrl });
+            var sql = "INSERT INTO sa_amna.Categories (Name, Description, ImageId) VALUES(@name, @description, @imageId)";
+            await connection.ExecuteAsync(sql, new { name = categoryFormData.Category.Name, description = categoryFormData.Category.Description, imageId = insertedImageId });
             _memoryCache.Remove(CacheKeys.AllCategories);
         }
 
@@ -90,12 +117,33 @@ namespace AmnasKitchen.Server.Services
         {
             using var connection = _databaseConnectionHandler.GetDbConnection();
             connection.Open();
+            var category = await connection.QuerySingleAsync<Category>("select * from sa_amna.Categories Where Id = @id", new { id = categoryId });
+
+            if (category == null)
+            {
+                throw new ArgumentNullException(nameof(category));
+            }
 
             var sql = "Delete from sa_amna.Categories Where Id = @id";
 
             await connection.ExecuteAsync(sql, new { id = categoryId });
 
+            if (category.ImageId != null)
+            {
+                await connection.ExecuteAsync("Delete from sa_amna.UploadedImages Where Id = @id", new { id = category.ImageId });
+            }
+
             _memoryCache.Remove(CacheKeys.AllCategories);
+        }
+
+        private async Task<int> StoreImageInDb(byte[] imageData)
+        {
+            using var connection = _databaseConnectionHandler.GetDbConnection();
+            connection.Open();
+
+            return await connection.QuerySingleAsync<int>(
+                       @"INSERT INTO sa_amna.UploadedImages (ImageData) VALUES (@imageData); SELECT CAST(SCOPE_IDENTITY() as int)",
+                       new { imageData });
         }
     }
 }
